Common AttachedPolicyCard
